requirements.txt
djlint==1.36.4
  click==8.3.1
  colorama==0.4.6
  cssbeautifier==1.15.4
    jsbeautifier==1.15.4
      six==1.17.0
      EditorConfig==0.17.1
    six==1.17.0
    EditorConfig==0.17.1
  jsbeautifier==1.15.4
    six==1.17.0
    EditorConfig==0.17.1
  json5==0.13.0
  pathspec==1.0.4
  PyYAML==6.0.3
  regex==2026.1.15
  tqdm==4.67.3
Flask==3.1.2
  blinker==1.9.0
  click==8.3.1
  itsdangerous==2.2.0
  Jinja2==3.1.6
    MarkupSafe==3.0.3
  MarkupSafe==3.0.3
  Werkzeug==3.1.5
    MarkupSafe==3.0.3
pipdeptree==2.30.0
  packaging==26.0
  pip==26.0.1
PyQt6-WebEngine==6.10.0
  PyQt6_sip==13.11.0
  PyQt6-WebEngine-Qt6==6.10.2
  PyQt6==6.10.2
    PyQt6_sip==13.11.0
    PyQt6-Qt6==6.10.2
pywebview==6.1
  proxy_tools==0.1.0
  bottle==0.13.4
  typing_extensions==4.15.0
QtPy==2.4.3
  packaging==26.0
